Split main.c into main.c, func.c, func.h
