+++ func.h
@@ -1,0 +1,13 @@
+#ifndef _HEADER_H_
+#define _HEADER_H_
+#include <stdio.h>
+#include <stdlib.h>
+
+void p1(int* p);
+void p2(int* p);
+void p3(int* p);
+void p4(int* p);
+
+int sclass(int a, int b);
+
+#endif
